application/requirements.txt
fastapi 
uvicorn 
GitPython
groq